requirements.txt
cachetools==5.3.1
colorlog==6.7.0
python-dotenv==1.0.0
Requests==2.31.0
SQLAlchemy==1.4.48
typer==0.9.0
